% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmodeling.R
\docType{package}
\name{statmodeling}
\alias{statmodeling}
\title{statmodeling: Statistical Modeling with R and Stan}
\description{
The \code{statmodeling} package provides a collection of
interactive \code{learnr} materials for the workshop
"Statistical Modeling with R and Stan". The workshop constists of six
sessions, each divided into a lecture and a lab. You can call the
materials for each lecture and lab using the code provided in the examples
below.
}
\examples{
\dontrun{
  # Introduction
  learnr::run_tutorial("00-int", package = "statmodeling")

  # Session 1
  learnr::run_tutorial("01-lec", package = "statmodeling")
  learnr::run_tutorial("01-lab", package = "statmodeling")

  # Session 2
  learnr::run_tutorial("02-lec", package = "statmodeling")
  learnr::run_tutorial("02-lab", package = "statmodeling")

  # etc
}
}
\seealso{
\code{\link{learnr}}
}
